Moved to reading values from variables initialized on startup instead of directly from EEPROM to decrease load and wear
Added a check to the scheduleManager to the loop to see what it's time to ring
This check is performed once every minute (calculated by taking a millis value minus the last ring time)

Overflow is prevented by using subtraction instead of addition, and an unsigned integer value so when subtracting it will produce the correct result even after the overflow occurs (50 days of up time)

Removed login page from the website, added logic to auth.js for logging in.

Authentication failure now results in a modal appearing on screen allowing you to log in. This move is hopefully more user friendly, as changes to a webpage aren't lost while logging in. As a bonus side effect it is also quite a bit faster.